+++ Controller/NotifyHandler.php
@@ -21,7 +21,7 @@ use Psr\Log\LoggerInterface;
 use Wirecard\Oxid\Core\PaymentMethodFactory;
 use Wirecard\Oxid\Core\ResponseHandler;
 use Wirecard\Oxid\Extend\Model\Order;
-use Wirecard\Oxid\Model\PaymentMethod;
+use Wirecard\Oxid\Model\PaymentMethod\PaymentMethod;
 use Wirecard\Oxid\Model\Transaction;
 
 use Wirecard\PaymentSdk\BackendService;
