+++ Core/OrderHelper.php
@@ -22,9 +22,10 @@ use Psr\Log\LoggerInterface;
 
 use Wirecard\Oxid\Extend\Model\Order;
 use Wirecard\Oxid\Extend\Model\Payment;
-use Wirecard\Oxid\Model\PaymentInAdvancePaymentMethod;
+use Wirecard\Oxid\Model\PaymentMethod\PaymentInAdvancePaymentMethod;
 use Wirecard\Oxid\Model\FormInteractionResponseFields;
 use Wirecard\Oxid\Model\PaymentInAdvancePaymentInformation;
+
 use Wirecard\PaymentSdk\BackendService;
 use Wirecard\PaymentSdk\Entity\Status;
 use Wirecard\PaymentSdk\Response\FailureResponse;
@@ -211,7 +212,7 @@ class OrderHelper
      */
     private static function _managePiaPaymentInformation($oResponse, $oOrder)
     {
-        if ($oOrder->oxorder__oxpaymenttype->value === PaymentInAdvancePaymentMethod::getName(true)) {
+        if ($oOrder->oxorder__oxpaymenttype->value === PaymentInAdvancePaymentMethod::getName()) {
             $oResponseXml = simplexml_load_string($oResponse->getRawData());
 
             $oSession = Registry::getSession();
