+++ Model/PaymentMethod/RatepayInvoicePaymentMethod.php
@@ -7,7 +7,7 @@
  * https://github.com/wirecard/oxid-ee/blob/master/LICENSE
  */
 
-namespace Wirecard\Oxid\Model;
+namespace Wirecard\Oxid\Model\PaymentMethod;
 
 use DateTime;
 use OxidEsales\Eshop\Core\Registry;
